pyproject.toml
[project]
name = "deep-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altair>=5.5.0",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.7",
    "kagglehub>=0.3.12",
    "matplotlib>=3.10.3",
    "numpy>=2.2.6",
    "pandas>=2.2.3",
    "polars>=1.30.0",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "torch>=2.7.0",
    "torchvision>=0.22.0",
    "transformers>=4.52.3",
]
